% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalyticsv2_operations.R
\name{add_application_cloud_watch_logging_option}
\alias{add_application_cloud_watch_logging_option}
\title{Adds an Amazon CloudWatch log stream to monitor application configuration errors}
\usage{
add_application_cloud_watch_logging_option(
  ApplicationName = "string",
  CurrentApplicationVersionId = 123,
  CloudWatchLoggingOption = list(
    LogStreamARN = "string"
  )
)
}
\arguments{
\item{ApplicationName}{[required] The Kinesis Data Analytics application name.}

\item{CurrentApplicationVersionId}{[required] The version ID of the Kinesis Data Analytics application. You can retrieve the application version ID using DescribeApplication.}

\item{CloudWatchLoggingOption}{[required] Provides the Amazon CloudWatch log stream Amazon Resource Name (ARN).}
}
\description{
Adds an Amazon CloudWatch log stream to monitor application configuration errors.
}
\examples{

}
