% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalyticsv2_operations.R
\name{create_application_snapshot}
\alias{create_application_snapshot}
\title{Creates a snapshot of the application's state data}
\usage{
create_application_snapshot(
  ApplicationName = "string",
  SnapshotName = "string"
)
}
\arguments{
\item{ApplicationName}{[required] The name of an existing application}

\item{SnapshotName}{[required] An identifier for the application snapshot.}
}
\description{
Creates a snapshot of the application's state data.
}
\examples{

}
