% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalyticsv2_operations.R
\name{delete_application_cloud_watch_logging_option}
\alias{delete_application_cloud_watch_logging_option}
\title{Deletes an Amazon CloudWatch log stream from an Amazon Kinesis Data Analytics application}
\usage{
delete_application_cloud_watch_logging_option(
  ApplicationName = "string",
  CurrentApplicationVersionId = 123,
  CloudWatchLoggingOptionId = "string"
)
}
\arguments{
\item{ApplicationName}{[required] The application name.}

\item{CurrentApplicationVersionId}{[required] The version ID of the application. You can retrieve the application version ID using DescribeApplication.}

\item{CloudWatchLoggingOptionId}{[required] The \code{CloudWatchLoggingOptionId} of the Amazon CloudWatch logging option to delete. You can get the \code{CloudWatchLoggingOptionId} by using the DescribeApplication operation.}
}
\description{
Deletes an Amazon CloudWatch log stream from an Amazon Kinesis Data Analytics application.
}
\examples{

}
