% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalyticsv2_operations.R
\name{describe_application}
\alias{describe_application}
\title{Returns information about a specific Amazon Kinesis Data Analytics application}
\usage{
describe_application(
  ApplicationName = "string",
  IncludeAdditionalDetails = TRUE|FALSE
)
}
\arguments{
\item{ApplicationName}{[required] The name of the application.}

\item{IncludeAdditionalDetails}{Displays verbose information about a Kinesis Data Analytics application, including the application's job plan.}
}
\description{
Returns information about a specific Amazon Kinesis Data Analytics application.
}
\details{
If you want to retrieve a list of all applications in your account, use the ListApplications operation.
}
\examples{

}
