% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalyticsv2_operations.R
\name{list_applications}
\alias{list_applications}
\title{Returns a list of Amazon Kinesis Data Analytics applications in your account}
\usage{
list_applications(Limit = NULL, NextToken = NULL)
}
\arguments{
\item{Limit}{The maximum number of applications to list.}

\item{NextToken}{If a previous command returned a pagination token, pass it into this value to retrieve the next set of results. For more information about pagination, see \href{https://docs.aws.amazon.com/cli/latest/userguide/pagination.html}{Using the AWS Command Line Interface's Pagination Options}.}
}
\description{
Returns a list of Amazon Kinesis Data Analytics applications in your account. For each application, the response includes the application name, Amazon Resource Name (ARN), and status.
}
\details{
If you want detailed information about a specific application, use DescribeApplication.
}
\section{Accepted Parameters}{
\preformatted{list_applications(
  Limit = 123,
  NextToken = "string"
)
}
}

