% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalyticsv2_package.R
\docType{package}
\name{paws.kinesisanalyticsv2-package}
\alias{paws.kinesisanalyticsv2}
\alias{paws.kinesisanalyticsv2-package}
\title{paws.kinesisanalyticsv2: Amazon Kinesis Analytics}
\description{
Documentation for Kinesis Data Analytics API v2
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
