% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalyticsv2_operations.R
\name{stop_application}
\alias{stop_application}
\title{Stops the application from processing data}
\usage{
stop_application(
  ApplicationName = "string"
)
}
\arguments{
\item{ApplicationName}{[required] The name of the running application to stop.}
}
\description{
Stops the application from processing data. You can stop an application only if it is in the running state. You can use the DescribeApplication operation to find the application state.
}
\examples{

}
